Fix NativeAOT official build (#64432)
Official builds are currently not building NativeAOT CoreLib.

For unknown reasons the official build splits native build and managed build of the CoreCLR partition.

We had a convenient clr.nativeaotlibs subset that built both the native part and managed part. Managed part can't be built without the native part, so it makes sense.

To satisfy official build's weirdness, we need to split this into two subsets so that we can tell the official build to build the managed part. (Official builds already build everything in the native part, so we're good there.)