Fix generic method resolution in composite R2R images (#64247)
According to my recent investigation, worse perf in the presence
of composite R2R images is due to the inability to resolve certain
generic methods in them; I believe the problem was caused by
inconsistent caching of the component assemblies. In particular,
DecodeModuleFromIndex uses the native metadata assembly ref cache
to resolve component assemblies but DecodeModuleFromIndexIfLoaded
doesn't consult it and direct load of an arbitrary assembly that
is a component to the composite image also doesn't set the cache
entry. This change fixes these deficiencies.

Thanks

Tomas