Check for multicast delegate before comparing invocation lists (#64729)
Fixes issue discovered in https://github.com/dotnet/runtime/pull/64404#discussion_r797220167.

The existing code worked but there is an off chance that a different kind of delegate (not multicast) could be storing a number in `m_extraFunctionPointerOrData` that matches the length of this delegate (length is stored in `m_extraFunctionPointerOrData`), making us fail with `InvalidCast`. We should only compare invocation lists once we know both delegates are multicast.