Fix SymbolixRegexRunner handling of "beginning" (#64946)
The Runner's Go implementation is loading the `runtext` but ignoring `runtextbeg`.  If a non-0 beginning were passed in, the matcher could end up evaluating beginning anchors incorrectly, as it always considers beginning == 0.  This fixes that by slicing the input to ensure the matcher is always called with a span representing the full beginning-end length, such that it's beginning==0 assumption is true.