Enable merging of platform-variant tests
This particular change doesn't yet attempt to merge platform-variant tests into
fewer source files mostly because the affected tests are mostly ILPROJ tests and
it's not yet clear to me how to apply platform variations to them without substantial
changes in the source code. For now I have just aligned the assembly vs. project
naming so that the tests work in the merged scenario.

Thanks

Tomas