Keep Obsolete and EditorBrowsable attributes consistent between ref and src and DefineConst clean-up (#65847)
* NET5_0_OR_GREATER, NET6_0_OR_GREATER -> NETCOREAPP

The minimum supported .NETCoreApp version in the repository is 6.0,
hence both defines can be replaced with the versionless NETCOREAPP one.
There is no need to version APIs below the minimum supported .NETCoreApp
version.

* Sync ObsoleteAttribute between ref and src

ApiCompat wasn't enabled to check if the ObsoleteAttribute is in sync
between the ref and the src assembly. Enabling it showed numerous
mismatches which this commit fixes.

Also making sure that the ApiCompat run that compares the live build
against the previous version of .NETCoreApp and .NETStandard2.x doesn't
complain about ObsoleteAttribute API changes as those are intentional.

* Sync EditorBrowsableAttribute between ref and src

The EditorBrowsableAttribute attribute wasn't enabled to be checked by
ApiCompat and in many cases there were mismatches. The most prominent
were InteropServices.

* HttpListenerContext pr feedback

* ResourceManager pr feedback

* Json PR feedback

* Fix cryptography PNSE build