Dedup patterm/options/timeout in Regex source generator (#66747)
* Dedup patterm/options/timeout in Regex source generator

If we end up with multiple identical pattern/options/timeout tuples in the same assembly, we don't need to emit unique code for each.  We can still emit the partial implementation for each, but emit a single implementation of the Regex they all share.

* Make the cached Regex members be static readonly fields

No reason we need a property here wrapping a compiler-generated backing field.